% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_to_cytoscape.R
\name{apply_custom_theme}
\alias{apply_custom_theme}
\title{apply_custom_theme}
\usage{
apply_custom_theme()
}
\description{
This function apply custom theme to current network in cytoscape. Here is the description of the theme.
}
\details{
\describe{
  \item{Fill color}{(based on type) gene = blue, protein = orange, GO = grey, side_effect = green, drug/compound = pink, pathway = red}
  \item{Node shape}{input diffusion = ELLIPSE, other = ROUND_RECTANGLE}
  \item{Node size}{(based on type) gene = 20, protein = 20, GO = 35, side_effect = 40, drug/compound = 40, pathway = 35.}
  \item{Node border}{gene_hepatox = red, other = black}
  }
}
\seealso{
export_to_cytoscape
}
