% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cons_Net.R
\name{Cons_Up_net}
\alias{Cons_Up_net}
\title{Construct ceRNA network that targets up-regulated mRNAs}
\usage{
Cons_Up_net(
  clipExpNum_mir2m_cutoff = 1,
  databaseNum_mir2m_cutoff = 3,
  clipExpNum_lnc2mir_cutoff = 1,
  clipExpNum_circ2mir_cutoff = 1
)
}
\arguments{
\item{clipExpNum_mir2m_cutoff}{the minimum CLIP-seq experiment number for miRNA-mRNA interactions (default = 1)}

\item{databaseNum_mir2m_cutoff}{the minimum number of databases supporting miRNA-mRNA interactions (default = 3)}

\item{clipExpNum_lnc2mir_cutoff}{the minimum CLIP-seq experiment number for lncRNA-miRNA interactions (default = 1)}

\item{clipExpNum_circ2mir_cutoff}{the minimum CLIP-seq experiment number for circRNA-miRNA interactions (default = 1)}
}
\value{
UPnetwork: ceRNA network that targets up-regulated mRNAs
}
\description{
Construct ceRNA network that targets up-regulated mRNAs
}
