% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_to_cytoscape.R
\name{export_to_cytoscape}
\alias{export_to_cytoscape}
\title{Export to Cytoscape}
\usage{
export_to_cytoscape(
  x,
  collectionName = "myNewCollection",
  networkName = "myNetwork"
)
}
\arguments{
\item{x}{igraph object, the network to export}

\item{collectionName}{character, name of collection inside Cytoscape}

\item{networkName}{character, name of network inside Cytoscape}
}
\description{
This function export an igraph object to cytoscape..
}
\examples{
\donttest{
data(liver_1.3_network)
data(liver_1.3_rwr_closest_dfr)
data(signature_maison)

signature_vids <- signature_maison$acetaminophen_all_all
res.diffusion <- get_route(liver_1.3_network, liver_1.3_rwr_closest_dfr, signature_vids, target_type = c("drug/compound", "pathway", "side_effect", "GO"))
x <- res.diffusion
export_to_cytoscape(res.diffusion)
apply_custom_theme()
}

}
