% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{liver_1.3_rwr_closest_dfr}
\alias{liver_1.3_rwr_closest_dfr}
\title{Closest Targets dfr}
\format{
data.frame with 4 columns: Seed, Target, Seed.type, Target.type
}
\usage{
liver_1.3_rwr_closest_dfr
}
\description{
Closest target from each node (each node is a seed), and for each layer.
}
\keyword{datasets}
