% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{module_ppi_M1_20231211}
\alias{module_ppi_M1_20231211}
\title{Module}
\format{
a data.frame
}
\usage{
module_ppi_M1_20231211
}
\description{
Precalulated PPI module using M1 olgo from MONET
}
\keyword{datasets}
