% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_network.R
\name{plot_highlighted_CEnetwork}
\alias{plot_highlighted_CEnetwork}
\title{Plot a ceRNA network with highlighted RNAs of interest}
\usage{
plot_highlighted_CEnetwork(
  network,
  highlight_node,
  highlight_circ_color = "#459943",
  highlight_lnc_color = "#F0E442",
  highlight_mir_color = "#CC79A7",
  highlight_edge_color = "#e41a1c",
  highlight_circ_alpha = 0.9,
  highlight_lnc_alpha = 0.9,
  highlight_mir_alpha = 0.9,
  highlight_m_alpha = 0.9,
  netlayout = NULL,
  pdf_height = 15,
  pdf_width = 15,
  vertex_label_family = "sans"
)
}
\arguments{
\item{network}{A graph object representing the ceRNA network, created using "plot_CEnetwork" before.}

\item{highlight_node}{Your RNA of interest}

\item{highlight_circ_color}{Color of highlighted circRNAs. Default is "#459943"}

\item{highlight_lnc_color}{Color of highlighted lncRNAs. Default is "#F0E442"}

\item{highlight_mir_color}{Color of highlighted miRNAs, Default is "#CC79A7"}

\item{highlight_edge_color}{Color of highlighted edges. Default is "#e41a1c"}

\item{highlight_circ_alpha}{a numeric value specifying the alpha transparency for highlighted circular RNA nodes}

\item{highlight_lnc_alpha}{a numeric value specifying the alpha transparency for highlighted long non-coding RNA nodes}

\item{highlight_mir_alpha}{a numeric value specifying the alpha transparency for highlighted microRNA nodes}

\item{highlight_m_alpha}{a numeric value specifying the alpha transparency for highlighted  mRNA nodes}

\item{netlayout}{A layout object representing the layout of the network. Default is NULL, and the function will calculate the layout using the Kamada-Kawai algorithm.}

\item{pdf_height}{An integer representing the height of the PDF output file in inches. Default is 15.}

\item{pdf_width}{An integer representing the width of the PDF output file in inches. Default is 15.}

\item{vertex_label_family}{A character string representing the font family for the vertex labels in the network plot. Default is "sans".}
}
\value{
This function outputs a PDF file of the network plot with the highlighted node and its connected edges colored differently.
}
\description{
This function plots a highlighted network with a specified node highlighted in a different color. The function takes as input the network and the node to be highlighted, and outputs a PDF file of the network plot with the highlighted node and its connected edges colored differently.
}
