% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{produce_diffusion_report}
\alias{produce_diffusion_report}
\title{Produce diffusion report}
\usage{
produce_diffusion_report(
  res_report,
  report_title,
  report_out_filepath = NA,
  overwrite = FALSE,
  render = FALSE
)
}
\arguments{
\item{res_report}{`report()` results}

\item{report_title}{(character) name of the report}

\item{report_out_filepath}{(character) filepath where to write the .Rmd report}

\item{overwrite}{(logical, default = FALSE) if TRUE, overwrite out filepath}

\item{render}{(logical, default = FALSE) if TRUE, render the html file}
}
\description{
Produce diffusion report based remarkable nodes in the network
}
\examples{
data(liver_1.3_network)
data(liver_1.3_rwr_closest_dfr)
data(signature_maison)
signature_vids <-  signature_maison$`acetaminophen_all_all`
res.diffusion <- get_route(liver_1.3_network, liver_1.3_rwr_closest_dfr, c(signature_vids), target_type = c("drug/compound", "pathway", "side_effect"))

res_report <- report(res.diffusion)
report_info_df <- produce_diffusion_report(res_report = res_report,  report_title = "acetaminophen", report_out_filepath = "report_acetaminophen_example.Rmd",
render = TRUE, overwrite = TRUE)

}
\seealso{
report
}
