% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEanalysis.R
\name{readexp}
\alias{readexp}
\title{Construct ceRNA network that targets your mRNA of interest}
\usage{
readexp(file, min_expr = 1, min_samples = 1)
}
\arguments{
\item{file}{a character string specifying the name of the input file to be read}

\item{min_expr}{a numeric value specifying the minimum expression level to consider a gene expressed (default: 1)}

\item{min_samples}{a numeric value specifying the minimum number of samples where a gene should be expressed (default: 1)}
}
\value{
a filtered gene expression data frame with genes as rows and samples as columns, where only genes with expression levels greater than or equal to min_expr in at least min_samples samples are retained. The filtered data frame is also written to a file named "<input_file_name>_filtered.txt".
}
\description{
Construct ceRNA network that targets your mRNA of interest (only support 1 mRNA, I will update later)
}
