% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{signature_maison}
\alias{signature_maison}
\title{Custom signature}
\format{
a list of char
}
\usage{
signature_maison
}
\description{
Custom gene signature from in house study.
}
\keyword{datasets}
