% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_RWR.R
\name{generate_closest_dfr}
\alias{generate_closest_dfr}
\title{Closest dfr}
\usage{
generate_closest_dfr(network, res_tmp_matrix)
}
\arguments{
\item{res_tmp_matrix}{}
}
\description{
From RWR result matrix generate closest dfr based on "type" attribue
}
