% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{report}
\alias{report}
\title{Report}
\usage{
report(x)
}
\arguments{
\item{x}{diffusion results from get_route()}
}
\value{
A list of data.frame:
\describe{
  \item{`degree_signature`}{degree of input nodes and hepatox infos}
  \item{`drugs`}{drug infos with DILI score, IC50}
  \item{`drugs.targets`}{drug to protein target}
  \item{`drugs.side_effect`}{drug side effects}
  \item{`drug.dist.signature`}{closest input node(s) for each drug and their distance}
  \item{`pathways`}{pathway infos}
  \item{`pathways.dist.signature`}{closest input node(s) for each pathway and their distance}
  \item{`GOs`}{GO infos}
  \item{`GO.dist.signature`}{closest input node(s) for each GO and their distance}
  }
}
\description{
Reporting remarkable nodes within the network. The function produce a list of data.frame containing the valuable informations.
}
\examples{
data(liver_1.3_network)
data(liver_1.3_rwr_closest_dfr)
data(signature_maison)
signature_vids <-  signature_maison$`acetaminophen_all_all`
res.diffusion <- get_route(liver_1.3_network, liver_1.3_rwr_closest_dfr, c(signature_vids, "DB00313"), target_type = c("drug/compound", "pathway"))
res_report <- report(res.diffusion)

}
