% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_RWR.R
\name{RWR_build_complete}
\alias{RWR_build_complete}
\title{` RWR scores}
\usage{
RWR_build_complete(X, verbose = TRUE)
}
\arguments{
\item{X}{(igraph) network}

\item{verbose}{(logical) default = TRUE}
}
\value{
a data.frame
}
\description{
Generate all random walk scores. Iteratively, each node become the seed.
Return the RWR for the connected nodes in the network as a squared data.frame.
}
